test: extract MockURLHandler.swift from UtilityTestSupport.swift
Splits `MockURLHandler` (URLOpening double) out of
UtilityTestSupport.swift so that file holds only the
`MockHotkeyManager` double.

Byte-identical move of lines 17-33. Both files retain their test imports.

UtilityTestSupport.swift: 53 → 36 lines. Tests: 667.
