+++ Tests/BugNarratorTests/UtilityTestSupport.swift
@@ -13,23 +13,6 @@ final class MockHotkeyManager: HotkeyManaging {
         registeredShortcuts.removeAll()
     }
 }
-
-final class MockURLHandler: URLOpening {
-    private(set) var openedURLs: [URL] = []
-    var shouldSucceed = true
-    var openResults: [Bool] = []
-
-    @discardableResult
-    func open(_ url: URL) -> Bool {
-        openedURLs.append(url)
-        if !openResults.isEmpty {
-            return openResults.removeFirst()
-        }
-
-        return shouldSucceed
-    }
-}
-
 @MainActor
 final class MockScreenCapturePermissionAccess: ScreenCapturePermissionAccessing {
     var permissionState: ScreenCapturePermissionState = .granted
